requirements.txt
streamlit>=1.20
pandas
requests
python-dotenv
groq 
